Fix copiler warning

+++ highs/mip/HighsRedcostFixing.cpp
@@ -292,28 +292,30 @@ void HighsRedcostFixing::addRootRedcost(const HighsMipSolver& mipsolver,
     double lb = mipsolver.mipdata_->getDomain().col_lower_[col];
     double ub = mipsolver.mipdata_->getDomain().col_upper_[col];
 
-    if (lpredcost[col] > mipsolver.mipdata_->feastol && lb > -kHighsIInf) {
+    if (lpredcost[col] > mipsolver.mipdata_->feastol &&
+        lb > -static_cast<double>(kHighsIInf)) {
       // col <= (cutoffbound - lpobj)/redcost + lb
       // so for lurkub = lb to ub - 1 we can compute the necessary cutoff
       // bound to reach this bound which is:
       //  lurkub = (cutoffbound - lpobj)/redcost + lb
       //  cutoffbound = (lurkub - lb) * redcost + lpobj
       findLurkingBounds(col, HighsInt{1}, static_cast<HighsInt>(lb),
-                        static_cast<HighsInt>(ub), ub < kHighsIInf, lpobjective,
+                        static_cast<HighsInt>(ub),
+                        ub < static_cast<double>(kHighsIInf), lpobjective,
                         lpredcost[col], maxNumSteps, maxNumStepsExp,
                         lurkingColUpper[col], lurkingColLower[col]);
     } else if (lpredcost[col] < -mipsolver.mipdata_->feastol &&
-               ub < kHighsIInf) {
+               ub < static_cast<double>(kHighsIInf)) {
       // col >= (cutoffbound - lpobj)/redcost + ub
       // so for lurklb = lb + 1 to ub we can compute the necessary cutoff
       // bound to reach this bound which is:
       //  lurklb = (cutoffbound - lpobj)/redcost + ub
       //  cutoffbound = (lurklb - ub) * redcost + lpobj
       findLurkingBounds(col, HighsInt{-1}, static_cast<HighsInt>(ub),
-                        static_cast<HighsInt>(lb), lb > -kHighsIInf,
-                        lpobjective, lpredcost[col], maxNumSteps,
-                        maxNumStepsExp, lurkingColLower[col],
-                        lurkingColUpper[col]);
+                        static_cast<HighsInt>(lb),
+                        lb > -static_cast<double>(kHighsIInf), lpobjective,
+                        lpredcost[col], maxNumSteps, maxNumStepsExp,
+                        lurkingColLower[col], lurkingColUpper[col]);
     }
   }
 }
